+++ indexer/SRIndexer.h
@@ -56,7 +56,7 @@ public:
 	void indexWithTrainedParams(arma::fmat& features);
 	void index(arma::fmat& features);
 
-    std::pair<vector<uint>,vector<float> > knnSearchIdLong(arma::fmat& name, int n, double search_limit);
+    std::pair<vector<unsigned long int>,vector<float> > knnSearchIdLong(arma::fmat& name, int n, double search_limit);
 
 	std::pair<vector<float>,vector<float> > knnSearchId(arma::fmat& name, int n, double search_limit);
 	std::pair<vector<float>,vector<float> > radiusSearchId(arma::fmat& name, double radius, int n, double search_limit);
